[OpenMP][CUDA] Fix the issue that P2P memcpy doesn't work
This patch fixes the issue that P2P memcpy doesn't work. The root cause is we didn't set current context when calling the API function.

This patch also adds a state to track if peer access is supported. It starts with `Unknown`. When a device is initialized, it checks the peer access to all other devices. If it is supported, set the state to `Yes`. Since the enabling requires to use the destination context, which may be `nullptr` when we initialize the source device, we add another state `Enabled`.

When we issue P2P memcpy, we only call the API if the state is `Enabled`. If the state is `Yes`, which means when we initialize source device, the destination device has not been initialized yet, we do the queries and set the state accordingly. If the state is `No`, we directly fall back to D2D memcpy. The state should not be `Unknown` at the moment.

Differential Revision: https://reviews.llvm.org/D122764