[OpenMP] Introduce kernel environment
This patch introduces per kernel environment. Previously, flags such as execution mode are set through global variables with name like `__kernel_name_exec_mode`. They are accessible on the host by reading the corresponding global variable, but not from the device. Besides, some assumptions, such as no nested parallelism, are not per kernel basis, preventing us applying per kernel optimization in the device runtime.

This is a combination and refinement of patch series D116908, D116909, and D116910.

Note: Tests have not been updated yet because there is a bug in `OpenMPOpt` that can break IR after optimization. The rest of the patches is ready for review. Tests will be updated after the bugs are fixed.

Differential Revision: https://reviews.llvm.org/D142569