Try to add support for rpc vprintf.
Try to add support for vprintf using rpc without implementing variadic arguments.
Using Buildtin to transform fprintf call to vfprintf & then replace vfprintf call to a call to __omp_vprintf.
This is the methode currently used to make printf work using rpcs.
fprintf is not recognise as a builtin for now, so it does not work.