Add "MPI mode" to OpenMP Device Runtime Library.
This is used to modify the number of available hardware threadi shown to the user to the size of the OpenMP team.
It is needed in order for OpenMP Applications to work on GPU with DeviceRTL MPI.
Also add getMaxTeamWarps that return the numbers of available active Warp/Wavefront in the current team/workgroup/threadblock (taking SPMD mode into account).